final_Cool_Art/requirements.txt
# Project Vibe - Hybrid Sketch Generator
# Python package requirements

# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Web Interface
gradio>=4.0.0

# Optional (for CUDA support)
# torch with CUDA: Install from https://pytorch.org/get-started/locally/
